minor refactor to get CI to pass
