--- packages/destination-actions/src/destinations/postscript/properties.ts
@@ -1,2 +1,0 @@
-// Postscript API Base URL
-export const PS_BASE_URL = 'https://api.postscript.io'
